Misc work on the "CreateMedia" and "CreateFolder" extension methods
Umbraco is quite picky when creating a new media without a parent. For instance, the a "MediaService.CreateMedia" method overload takes "IMedia" as the parent, but the method will throw an exception if the specified parent is null - even though the "parent" parameter is marked as nullable. Another overload takes a GUID as the parent key, but the parameter is not nullable, and specifying Guid.Empty also triggers an exception.

- fixed misc issues with speciyfing a nullable parent
- added misc new method overloads
- etc.